refactor(stripe): query-ify getOrganization
