+++ jsapp/js/account/plans/planButton.component.tsx
@@ -1,13 +1,8 @@
 import BillingButton from 'js/account/plans/billingButton.component';
 import React, {useContext} from 'react';
-import type {
-  Price,
-  Organization,
-  SinglePricedProduct,
-} from 'js/account/stripe.types';
-import {postCustomerPortal} from 'js/account/stripe.api';
+import type {Price, SinglePricedProduct} from 'js/account/stripe.types';
+import {postCustomerPortal, useOrganizationQuery} from 'js/account/stripe.api';
 import {processCheckoutResponse} from 'js/account/stripe.utils';
-import {useOrganizationQuery} from 'js/query/queries/organizations.query';
 
 interface PlanButtonProps {
   buySubscription: (price: Price, quantity?: number) => void;
